Working code commit

+++ crawl.py
@@ -7,6 +7,20 @@ allowed_mimes = storage_profiles['allowed_mimes']
 s3_bucket = storage_profiles['s3_bucket']
 still_image_storage_path = storage_profiles['video_still']
 
+# test for ffmpeg if video/mp4 is in mime config
+if 'video/mp4' in allowed_mimes:
+    ffmpeg_path = subprocess.run(['which', 'ffmpeg'], stdout=subprocess.PIPE).stdout.decode('utf-8')
+    if not ffmpeg_path:
+        print("no ffmpeg installed, can't process video/mp4!")
+        sys.exit(0)
+
+# make sure if uploading to s3 that aws-cli (aws) is installed
+if not storage_profiles['s3_bucket']:
+    aws_cli_path = subprocess.run(['which', 'aws'], stdout=subprocess.PIPE).stdout.decode('utf-8')
+    if not aws_cli_path:
+        print("no aws-cli installed, can't upload to s3!")
+        sys.exit(0)
+
 for site_key, pull_list in site_block.items():
     if 'reddit' in site_key:
         allowed_sources = reddit_profile['allowed_sources']
@@ -21,9 +35,10 @@ for site_key, pull_list in site_block.items():
         print('Processing Reddit...')
         for pull_source in pull_list:
             print('\t/r/' + pull_source[0])
-            for submission in reddit.subreddit(pull_source[0]).hot(limit=25):
+            for submission in reddit.subreddit(pull_source[0]).hot(limit=reddit_profile['item_count']):
                 site_tag = pull_source[1]
                 # check if submission domain is allowed. this is reddit specific.
+                print('trying' + submission.url)
                 if submission.domain in allowed_sources:
                     object_url = submission.url
                     object_title = submission.title
@@ -51,15 +66,16 @@ for site_key, pull_list in site_block.items():
                             uploadToS3(object_storage_path, object_final_reference, object_mime, s3_bucket)
                             # store details in database
                             storeObjectDetails(object_title, object_final_reference, storage_type, site_tag)
-                            try:
-                                print('clearing files...')
-                                os.remove(object_storage_path + object_final_reference)
-                                os.remove(still_image_storage_path + object_final_reference.replace('mp4', 'jpeg'))
-                            except OSError as e:
-                                print(e)
-                            print('did one... sleeping for ' + str(sleep_seconds) + ' seconds...')
-        if sleep_seconds:
-            time.sleep(sleep_seconds)
+                            if not keep_local_files:
+                                try:
+                                    print('clearing files...')
+                                    os.remove(object_storage_path + object_final_reference)
+                                    os.remove(still_image_storage_path + object_final_reference.replace('mp4', 'jpeg'))
+                                except OSError as e:
+                                    print(e)
+            if sleep_seconds:
+                print('sleeping ' + str(sleep_seconds) + 's')
+                time.sleep(sleep_seconds)
     elif 'yourkeyhere':
         # here is probably a good spot for site specific handling rules
         pass
